added a new js file with the interaction functions
