+++ src/modules/taskManager.js
@@ -4,7 +4,7 @@ export function addTask(task) {
   tasks.push(task);
 }
 
-function updateIndexes() {
+export function updateIndexes() {
   tasks.forEach((task, index) => {
     task.index = index + 1;
   });
@@ -20,18 +20,6 @@ export function editTaskDescription(taskId, newDescription) {
   if (task) {
     task.description = newDescription;
   }
-}
-
-export function toggleTaskCompleted(taskId) {
-  const task = tasks.find((task) => task.index === taskId);
-  if (task) {
-    task.completed = !task.completed;
-  }
-}
-
-export function clearCompletedTasks() {
-  tasks = tasks.filter((task) => !task.completed);
-  updateIndexes();
 }
 
 export function getTasks() {
